removed silly uses of 'long' (now int, or qbyte *, depending on the way it was used), and changed memheader sentinels to work on processors which can not handle unaligned memory access (Sparc for example)
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@1734 d7cf8633-e32d-0410-b094-e92efae38249

+++ gl_rsurf.c
@@ -579,7 +579,7 @@ static void RSurfShader_Sky(msurface_t *firstsurf)
 				#if SLOWMATH
 				length = 3.0f / sqrt(number);
 				#else
-				*((long *)&length) = 0x5f3759df - ((* (long *) &number) >> 1);
+				*((int *)&length) = 0x5f3759df - ((* (int *) &number) >> 1);
 				length = 3.0f * (length * (1.5f - (number * 0.5f * length * length)));
 				#endif
 
@@ -653,7 +653,7 @@ static void RSurfShader_Sky(msurface_t *firstsurf)
 				#if SLOWMATH
 				length = 3.0f / sqrt(number);
 				#else
-				*((long *)&length) = 0x5f3759df - ((* (long *) &number) >> 1);
+				*((int *)&length) = 0x5f3759df - ((* (int *) &number) >> 1);
 				length = 3.0f * (length * (1.5f - (number * 0.5f * length * length)));
 				#endif
